ti964: map subdev alias i2c address
Each sensor device plugged into ti964 has 2 i2c address,
the physical address and the alias address which need
to set to ti964 registers. This will remove i2c address
dependency in ti964.

ti964 will get write/read command for each alias address
and map to its physical address.

+++ include/media/ti964.h
@@ -50,7 +50,7 @@ struct ti964_csi_data_format {
 struct ti964_subdev_info {
 	struct i2c_board_info board_info;
 	int i2c_adapter_id;
-	unsigned int rx_port;
+	unsigned short rx_port;
 	unsigned short phy_i2c_addr;
 };
